pyproject.toml
[tool.poetry]
name = "canalservice"
version = "0.1.0"
description = "testwork"
authors = ["Aleksey Lesnikov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.82.0"
oauth2client = "^4.1.3"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.5"
apscheduler = "^3.10.1"
asyncio = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
